+++ src/Presentation/SmartStore.Web/Controllers/CatalogHelper.cs
@@ -285,6 +285,92 @@ namespace SmartStore.Web.Controllers
 			return model;
 		}
 
+		public void SelectProductAttributeValues(ProductDetailsModel model, NameValueCollection queryData)
+		{
+			Guard.ArgumentNotNull(() => model);
+
+			if (queryData == null || queryData.Count <= 0)
+				return;
+
+			try
+			{
+				var items = queryData.AllKeys
+					.Where(x => x.EmptyNull().StartsWith("product_attribute_"))
+					.SelectMany(queryData.GetValues, (k, v) => new { key = k, value = v });
+
+				foreach (var item in items)
+				{
+					var ids = item.key.Replace("product_attribute_", "").SplitSafe("_");
+					if (ids.Count() > 3)
+					{
+						var productId = ids[0].ToInt();
+						var bundleItemId = ids[1].ToInt();
+						var productAttributeId = ids[2].ToInt();
+						//var attributeId = ids[3].ToInt();
+
+						var attribute = model.ProductVariantAttributes.FirstOrDefault(x =>
+							x.ProductId == productId &&
+							x.BundleItemId == bundleItemId &&
+							x.ProductAttributeId == productAttributeId);
+
+						if (attribute != null)
+						{
+							switch (attribute.AttributeControlType)
+							{
+								case AttributeControlType.DropdownList:
+								case AttributeControlType.RadioList:
+								case AttributeControlType.Checkboxes:
+								case AttributeControlType.ColorSquares:
+									var attributeValue = attribute.Values.FirstOrDefault(x => x.Name.IsCaseInsensitiveEqual(item.value));
+									if (attributeValue != null)
+									{
+										attribute.Values.Each(x => x.IsPreSelected = false);
+										attributeValue.IsPreSelected = true;
+									}
+									break;
+								case AttributeControlType.TextBox:
+								case AttributeControlType.MultilineTextbox:
+									attribute.TextValue = item.value;
+									break;
+								case AttributeControlType.Datepicker:
+									if (item.value.Length == 8)
+									{
+										attribute.SelectedYear = item.value.Substring(0, 4).ToInt();
+										attribute.SelectedMonth = item.value.Substring(4, 2).ToInt();
+										attribute.SelectedDay = item.value.Substring(6, 2).ToInt();
+									}
+									break;
+							}
+						}
+					}
+				}
+
+				items = queryData.AllKeys
+					.Where(x => x.EmptyNull().StartsWith("giftcard_"))
+					.SelectMany(queryData.GetValues, (k, v) => new { key = k, value = v });
+
+				foreach (var item in items)
+				{
+					var key = item.key.EmptyNull().ToLower();
+
+					if (key.EndsWith("recipientname"))
+						model.GiftCard.RecipientName = item.value;
+					else if (key.EndsWith("recipientemail"))
+						model.GiftCard.RecipientEmail = item.value;
+					else if (key.EndsWith("sendername"))
+						model.GiftCard.SenderName = item.value;
+					else if (key.EndsWith("senderemail"))
+						model.GiftCard.SenderEmail = item.value;
+					else if (key.EndsWith("message"))
+						model.GiftCard.Message = item.value;
+				}
+			}
+			catch (Exception exception)
+			{
+				Logger.Warning(exception.ToAllMessages(), exception);
+			}
+		}
+
 		public void PrepareProductReviewsModel(ProductReviewsModel model, Product product)
 		{
 			if (product == null)